Split the large circleci image into multiple smaller ones
A full image can take up to 6 minutes to setup in CicleCI when there
is not cache on the dispatched container. So split the image to the
minimum necessary for each build. This reduces the initial 15GB image
size to:

lkldocker/circleci-x86_64          493MB
lkldocker/circleci-mingw           1.45GB
lkldocker/circleci-android-arm64   9.56GB
lkldocker/circleci-android-arm32   9.49GB
lkldocker/circleci-android         4.85GB
lkldocker/circleci                 420MB

Also added makefile to test build the change locally. This is not used
to publish the images, that happens via docker hub at
https://hub.docker.com/r/lkldocker/.

Signed-off-by: Octavian Purdila <tavi@cs.pub.ro>