Add tag tracking in Makefile
This commit introduces a process of commiting tags. Pushing certain
tags kicks the docker hub build but they need to properly formatted
and be accompanied by commits to work.

To publish a new tag and trigger a docker hub build modify the
TAG_ variables in the makefile then issue:

 $ make tag-<image>

to generate the actual git tag.

This process allows us avoid typos in the tag names as well allowing
us to quickly know what are the current tags. It also makes it easy to
track the tag updates in the commit log.

Signed-off-by: Octavian Purdila <tavi@cs.pub.ro>