use circleci instead of docker hub autobuild
This commit introduces autobuild triggered by pull requests and push to
master branch so that we can automate the test before uploading docker
hub.

The workflow to deploy docker images is:

- make changes to master branch, or merge a pull request to master
  branch
- (manually) tag to the commit on the master branch

This tag will trigger an autobuild on circleci and upload images to
docker hub.  Pull requests only trigger build without the deployment.

Signed-off-by: Hajime Tazaki <thehajime@gmail.com>